pyproject.toml
[tool.poetry]
name = "chemical-treatment"
version = "0.1.0"
description = "Chemical structure rendering service"
authors = ["DPavlen"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.15"
drf-spectacular = "^0.29"
django-configurations = "^2.5.1"
python-dotenv = "^1.2.1"
epam-indigo = "^1.33"
pillow = "^11.0"
djangorestframework-simplejwt = "^5.3"
psycopg2-binary = "^2.9"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
isort = "^5.13.0"

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
